+++ frontend/app/components/home/Card.vue
